[GYJB-157] CorsConfig Class 분리
